+++ examples/simulation/plot_simulate_raw_data.py
@@ -3,8 +3,8 @@
 Generate simulated raw data
 ===========================
 
-This example generates raw data by repeating a desired source
-activation multiple times.
+This example generates raw data by repeating a desired source activation
+multiple times.
 """
 # Authors: Yousra Bekhti <yousra.bekhti@gmail.com>
 #          Mark Wronkiewicz <wronk.mark@gmail.com>