+++ mne/simulation/evoked.py
@@ -189,7 +189,8 @@ def _generate_noise(info, cov, iir_filter, random_state, n_samples, zi=None):
     """Create spatially colored and temporally IIR-filtered noise."""
     from scipy.signal import lfilter
     rng = check_random_state(random_state)
-    _, _, colorer = compute_whitener(cov, info, pca=True, return_colorer=True)
+    _, _, colorer = compute_whitener(cov, info, pca=True, return_colorer=True,
+                                     verbose=False)
     noise = np.dot(colorer, rng.randn(colorer.shape[1], n_samples))
     if iir_filter is not None:
         if zi is None: